refactor(warehouse): store stock quantities as integers
Qty columns move from decimal(15,4) to integer so 10 stays 10 instead
of 10.0000. Covers transfer/request/adjustment items, layers,
movements and discrepancies; unit_cost stays decimal. Validations and
form inputs enforce whole units.

+++ Modules/Warehouse/app/Http/Requests/ApproveStockRequestRequest.php
@@ -40,7 +40,7 @@ class ApproveStockRequestRequest extends FormRequest
         return [
             'items' => ['required', 'array', 'min:1'],
             'items.*.product_variant_id' => ['required', 'integer', 'exists:product_variants,id'],
-            'items.*.qty_approved' => ['required', 'numeric', 'min:0'],
+            'items.*.qty_approved' => ['required', 'integer', 'min:0'],
         ];
     }
 }
